+++ pyagentspec/src/pyagentspec/tools/__init__.py
@@ -8,6 +8,12 @@
 
 from .builtintool import BuiltinTool
 from .clienttool import ClientTool
+from .codeexecutors import (
+    CodeExecutor,
+    EndpointCodeExecutor,
+    LocalContainerCodeExecutor,
+    SubProcessCodeExecutor,
+)
 from .remotetool import RemoteTool
 from .servertool import ServerTool
 from .tool import Tool
@@ -15,6 +21,10 @@ from .toolbox import ToolBox
 
 __all__ = [
     "ClientTool",
+    "CodeExecutor",
+    "SubProcessCodeExecutor",
+    "LocalContainerCodeExecutor",
+    "EndpointCodeExecutor",
     "ServerTool",
     "BuiltinTool",
     "RemoteTool",